requirements/requirements_heroku.txt
django
django_heroku
gunicorn
mysqlclient
pandas
python-dateutil
urllib3